src/diploma_sar_detection/requirements-gpu-cu124.txt
--index-url https://download.pytorch.org/whl/cu124
--extra-index-url https://pypi.org/simple

torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
ultralytics
timm
pyyaml
matplotlib
pandas
tqdm
onnxruntime
onnx
onnxslim
numpy
pillow
